refactor: Streamline map_processor.py
* **Modular Architecture**: Decomposed the monolithic map processing system into specialized modules using a clean architectural approach
  * Created dedicated action handler modules for markers, regions, geometry, data and view operations
  * Extracted specialized services for weather data processing and risk analysis

* **Design Pattern Implementation**:
  * Implemented the Strategy Pattern with function-based handlers for different map operations
  * Added a Factory Method approach with the `get_action_handlers()` registry for dynamic handler lookup
  * Applied Decorator Pattern with `create_handler` to standardize error handling across all handlers

* **Code Organization Improvements**:
  * Standardized interface for action handlers with consistent parameter and return types
  * Created a consistent bounds management system across all actions
  * Implemented proper exception handling throughout the system

* **Enhanced Geographic Data Processing**:
  * Improved location filtering and matching for weather data
  * Added sophisticated coordinate buffer calculations for location-based queries
  * Created standardized GeoJSON serialization to prevent timestamp conversion issues

* **Documentation Enhancements**:
  * Added comprehensive docstrings explaining function purposes, parameters, and return values
  * Included clarifying comments for complex operations like coordinate transformations and geographic calculations
  * Documented edge cases and special handling for different geographic data types

* **Maintainability Improvements**:
  * Reduced average function size from 100+ lines to under 30 lines
  * Eliminated deeply nested conditional logic
  * Preserved the original implementation in `map_processor_BEFORE_REFACTOR.py` for reference
